HS-30 RENAME: Divide functions in new file
